// src/types/data/movieCredits.ts
export type MovieCastProps = {
    adult: boolean
    gender: number
    id: number
    known_for_department: string
    name: string
    original_name: string
    popularity: number
    profile_path: string | null
    cast_id: number
    character: string
    credit_id: string
    order: number
}

export type MovieCrewProps = {
    adult: boolean
    gender: number
    id: number
    known_for_department: string
    name: string
    original_name: string
    popularity: number
    profile_path: string | null
    credit_id: string
    department: string
    job: string
}

export type MovieCreditsProps = {
    id: number
    cast: MovieCastProps[]
    crew: MovieCrewProps[]
}